Replace yarn 1 with pnpm (#19870)
Due to the end of support for Yarn 1 and several bugs that were
impacting our workflows, we realized a need to update the package
manager. After discussion and analysis (p9dueE-2MY-p2) we decided to go
with pnpm as the replacement.

See docs/yarn-upgrade.md for further notes.

Other side effects of this PR:

* Update a bunch of indirect dependencies.
* Add some missing dependencies.
* Move the js-test-runner into its own "project", like the Jetpack CLI,
  so we can remove its deps from the monorepo-root package.json.
* Deprecated a bunch of scripts in the monorepo-root package.json.
* Our CI now uses the published version of action-required-review.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/896884684