Remove unused JS deps (#31881)
* js-packages/storybook: Remove unused deps on `@emotion/*`

Uses of all except `@emotion/babel-plugin` were removed in #30015. And
the babel plugin doesn't seem to do anything here, build is unchanged
with it removed.

* Remove all deps on `@babel/register`

It only takes effect if you require it in a script somewhere, which we
haven't done anywhere since js-test-runner was removed in #24830.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5552854849